Multiplication draft
This will look like a complete mess.  But I wanted to get this out there
now so that we can iterate on the details.

I need to look into how we manage subscripts as a high priority item.
There are a bunch of places where we should be tagging stuff as code or
using pseudocode blocks, but I stopped doing the conversion for now
until I sort out the subscript business.